backend/requirements.txt
# Backend Dependencies - Delirium Infinite
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
aiofiles==24.1.0
redis==5.2.1
qdrant-client==1.12.1
openai==1.58.1
anthropic==0.42.0
google-generativeai==0.8.4
python-dotenv==1.0.1
jinja2==3.1.5
aiosqlite==0.20.0
psutil==6.1.1
